refactor(arch): split agent.py and hybrid.py into focused modules
Split src/orchestration/agent.py (911 lines) into:
- agent_loop.py: Core ReAct loop, tools, public agent_query()
- complexity.py: Question complexity detection
- decomposition.py: Question decomposition + synthesis
- voting.py: Multi-trajectory parallel execution + majority voting
- _parsing.py: Shared JSON response parsing

Split src/retrieval/hybrid.py (702 lines) into:
- bm25.py: BM25 fulltext search
- vector.py: Dense vector similarity (legacy + Cypher 25)
- graph.py: Entity-based graph search + multi-hop expansion
- community.py: Community embedding retrieval
- fusion.py: WRRF fusion + orchestration + query_graph

Public interfaces unchanged via re-export shims.
Added backward-compat shims: src/agent.py, src/retrieve.py,
src/reranker.py, src/neo4j_client.py.

Tests updated to patch at correct module paths.
70 tests passing, 0 lint errors.
