+++ src/neo4j_client.py
@@ -1,0 +1,3 @@
+"""Backward-compatibility shim: imports redirect to src.infrastructure.neo4j_client."""
+
+from src.infrastructure.neo4j_client import neo4j_client  # noqa: F401
