+++ src/orchestration/__init__.py
@@ -1,0 +1,4 @@
+"""Orchestration package — re-exports for backward compatibility."""
+
+from src.orchestration.agent_loop import agent_query  # noqa: F401
+from src.orchestration.voting import run_agent_scaled  # noqa: F401
