+++ src/orchestration/_parsing.py
@@ -1,0 +1,24 @@
+"""Shared utilities for parsing agent JSON responses."""
+
+from __future__ import annotations
+
+import json
+import re
+
+
+def parse_agent_response(text: str) -> dict | None:
+    """Parse JSON from model output, handling markdown fences."""
+    cleaned = text.strip()
+    if cleaned.startswith("```"):
+        cleaned = re.sub(r"^```[a-zA-Z]*\n?", "", cleaned)
+        cleaned = re.sub(r"\n?```$", "", cleaned)
+        cleaned = cleaned.strip()
+
+    match = re.search(r"\{.*\}", cleaned, re.DOTALL)
+    if not match:
+        return None
+
+    try:
+        return json.loads(match.group())
+    except json.JSONDecodeError:
+        return None
