+++ src/reranker.py
@@ -1,0 +1,3 @@
+"""Backward-compatibility shim: imports redirect to src.retrieval.reranker."""
+
+from src.retrieval.reranker import rerank  # noqa: F401
