+++ src/retrieval/__init__.py
@@ -1,0 +1,3 @@
+"""Retrieval package — re-exports for backward compatibility."""
+
+from src.retrieval.fusion import QueryResult, hybrid_retrieve, query_graph  # noqa: F401
