+++ tests/test_hybrid_retrieve.py
@@ -1,12 +1,15 @@
-"""Tests for hybrid retrieval functions in src/retrieve.py."""
+"""Tests for hybrid retrieval functions."""
 
 from __future__ import annotations
 
 from contextlib import contextmanager
 
-
-import src.retrieve as retrieve_mod
-from src.retrieve import _wrrf_fuse, _vector_search, _graph_search, hybrid_retrieve
+import src.retrieval.fusion as fusion_mod
+import src.retrieval.vector as vector_mod
+import src.retrieval.graph as graph_mod
+from src.retrieval.fusion import _wrrf_fuse, hybrid_retrieve
+from src.retrieval.vector import vector_search as _vector_search
+from src.retrieval.graph import graph_search as _graph_search
 
 
 # ---------------------------------------------------------------------------
@@ -40,12 +43,27 @@ class _FakeSession:
         pass
 
 
-def _make_fake_session_factory(results: list[dict]):
-    @contextmanager
-    def _fake_session():
-        yield _FakeSession(results)
+class _FakeNeo4jClient:
+    """Fake neo4j_client that returns a context-managed session."""
 
-    return _fake_session
+    def __init__(self, results: list[dict]):
+        self._results = results
+
+    @contextmanager
+    def session(self):
+        yield _FakeSession(self._results)
+
+
+class _FailingNeo4jClient:
+    """Fake neo4j_client whose session() always raises."""
+
+    def __init__(self, exc: Exception):
+        self._exc = exc
+
+    @contextmanager
+    def session(self):
+        raise self._exc
+        yield  # noqa: RET503
 
 
 # ---------------------------------------------------------------------------
@@ -58,9 +76,9 @@ class TestWrrfFuse:
 
     def test_basic_fusion_three_channels(self, monkeypatch) -> None:
         """Verify WRRF formula with 3 channels and known rankings."""
-        monkeypatch.setattr(retrieve_mod.settings, "wrrf_weight_bm25", 0.4)
-        monkeypatch.setattr(retrieve_mod.settings, "wrrf_weight_vector", 0.4)
-        monkeypatch.setattr(retrieve_mod.settings, "wrrf_weight_graph", 0.2)
+        monkeypatch.setattr(fusion_mod.settings, "wrrf_weight_bm25", 0.4)
+        monkeypatch.setattr(fusion_mod.settings, "wrrf_weight_vector", 0.4)
+        monkeypatch.setattr(fusion_mod.settings, "wrrf_weight_graph", 0.2)
 
         bm25 = [_make_row("c1", "P1"), _make_row("c2", "P2")]
         vector = [_make_row("c2", "P2"), _make_row("c3", "P3")]
@@ -73,11 +91,8 @@ class TestWrrfFuse:
 
         scores = {r["chunk_id"]: r["score"] for r in results}
 
-        # c1: bm25 rank=1, graph rank=1 -> 0.4/(60+1) + 0.2/(60+1)
         expected_c1 = 0.4 / (60 + 1) + 0.2 / (60 + 1)
-        # c2: bm25 rank=2, vector rank=1 -> 0.4/(60+2) + 0.4/(60+1)
         expected_c2 = 0.4 / (60 + 2) + 0.4 / (60 + 1)
-        # c3: vector rank=2, graph rank=2 -> 0.4/(60+2) + 0.2/(60+2)
         expected_c3 = 0.4 / (60 + 2) + 0.2 / (60 + 2)
 
         assert abs(scores["c1"] - expected_c1) < 1e-9
@@ -85,9 +100,8 @@ class TestWrrfFuse:
         assert abs(scores["c3"] - expected_c3) < 1e-9
 
     def test_includes_channel_scores(self, monkeypatch) -> None:
-        """Each result should have channel_scores showing per-channel contribution."""
-        monkeypatch.setattr(retrieve_mod.settings, "wrrf_weight_bm25", 0.5)
-        monkeypatch.setattr(retrieve_mod.settings, "wrrf_weight_vector", 0.5)
+        monkeypatch.setattr(fusion_mod.settings, "wrrf_weight_bm25", 0.5)
+        monkeypatch.setattr(fusion_mod.settings, "wrrf_weight_vector", 0.5)
 
         results_by_channel = {
             "bm25": [_make_row("c1")],
@@ -102,27 +116,26 @@ class TestWrrfFuse:
         assert "vector" in results[0]["channel_scores"]
 
     def test_empty_channels_returns_empty(self, monkeypatch) -> None:
-        monkeypatch.setattr(retrieve_mod.settings, "wrrf_weight_bm25", 0.4)
-        monkeypatch.setattr(retrieve_mod.settings, "wrrf_weight_vector", 0.4)
+        monkeypatch.setattr(fusion_mod.settings, "wrrf_weight_bm25", 0.4)
+        monkeypatch.setattr(fusion_mod.settings, "wrrf_weight_vector", 0.4)
 
         results = _wrrf_fuse({"bm25": [], "vector": []}, k=60)
         assert results == []
 
     def test_single_channel_nonempty(self, monkeypatch) -> None:
-        monkeypatch.setattr(retrieve_mod.settings, "wrrf_weight_bm25", 0.4)
+        monkeypatch.setattr(fusion_mod.settings, "wrrf_weight_bm25", 0.4)
 
         bm25 = [_make_row("c1"), _make_row("c2"), _make_row("c3")]
         results = _wrrf_fuse({"bm25": bm25}, k=60)
 
         assert len(results) == 3
-        # Order preserved from single channel
         assert results[0]["chunk_id"] == "c1"
         assert results[1]["chunk_id"] == "c2"
         assert results[2]["chunk_id"] == "c3"
 
     def test_deduplication(self, monkeypatch) -> None:
-        monkeypatch.setattr(retrieve_mod.settings, "wrrf_weight_bm25", 0.4)
-        monkeypatch.setattr(retrieve_mod.settings, "wrrf_weight_vector", 0.4)
+        monkeypatch.setattr(fusion_mod.settings, "wrrf_weight_bm25", 0.4)
+        monkeypatch.setattr(fusion_mod.settings, "wrrf_weight_vector", 0.4)
 
         bm25 = [_make_row("c1", "PageA"), _make_row("c2", "PageB")]
         vector = [_make_row("c1", "PageA_vec"), _make_row("c2", "PageB_vec")]
@@ -133,9 +146,8 @@ class TestWrrfFuse:
         assert len(chunk_ids) == len(set(chunk_ids))
 
     def test_community_channel_weight(self, monkeypatch) -> None:
-        """Community channel uses its own weight from settings."""
-        monkeypatch.setattr(retrieve_mod.settings, "wrrf_weight_bm25", 0.0)
-        monkeypatch.setattr(retrieve_mod.settings, "wrrf_weight_community", 1.0)
+        monkeypatch.setattr(fusion_mod.settings, "wrrf_weight_bm25", 0.0)
+        monkeypatch.setattr(fusion_mod.settings, "wrrf_weight_community", 1.0)
 
         results_by_channel = {
             "bm25": [_make_row("c1")],
@@ -145,13 +157,12 @@ class TestWrrfFuse:
         results = _wrrf_fuse(results_by_channel, k=60)
 
         scores = {r["chunk_id"]: r["score"] for r in results}
-        # c2 from community with weight 1.0 should score higher than c1 from bm25 with weight 0.0
         assert scores["c2"] > scores["c1"]
 
     def test_results_sorted_descending(self, monkeypatch) -> None:
-        monkeypatch.setattr(retrieve_mod.settings, "wrrf_weight_bm25", 0.4)
-        monkeypatch.setattr(retrieve_mod.settings, "wrrf_weight_vector", 0.4)
-        monkeypatch.setattr(retrieve_mod.settings, "wrrf_weight_graph", 0.2)
+        monkeypatch.setattr(fusion_mod.settings, "wrrf_weight_bm25", 0.4)
+        monkeypatch.setattr(fusion_mod.settings, "wrrf_weight_vector", 0.4)
+        monkeypatch.setattr(fusion_mod.settings, "wrrf_weight_graph", 0.2)
 
         bm25 = [_make_row("c1"), _make_row("c2"), _make_row("c3")]
         vector = [_make_row("c3"), _make_row("c1"), _make_row("c2")]
@@ -170,14 +181,13 @@ class TestWrrfFuse:
 
 class TestVectorSearch:
     def test_returns_results_with_legacy_cypher(self, monkeypatch) -> None:
-        monkeypatch.setattr(retrieve_mod.settings, "neo4j_use_search_clause", False)
+        monkeypatch.setattr(vector_mod.settings, "neo4j_use_search_clause", False)
 
         fake_rows = [
             {"page_title": "P1", "page_url": "http://p1", "page_id": "pid1",
              "chunk_id": "c1", "chunk_text": "text1", "vector_score": 0.95},
         ]
-        fake = _make_fake_session_factory(fake_rows)
-        monkeypatch.setattr(retrieve_mod.neo4j_client, "session", fake)
+        monkeypatch.setattr(vector_mod, "neo4j_client", _FakeNeo4jClient(fake_rows))
 
         results = _vector_search([0.1] * 1024, top_k=5)
 
@@ -186,14 +196,13 @@ class TestVectorSearch:
         assert results[0]["vector_score"] == 0.95
 
     def test_returns_results_with_search_clause(self, monkeypatch) -> None:
-        monkeypatch.setattr(retrieve_mod.settings, "neo4j_use_search_clause", True)
+        monkeypatch.setattr(vector_mod.settings, "neo4j_use_search_clause", True)
 
         fake_rows = [
             {"page_title": "P2", "page_url": "http://p2", "page_id": "pid2",
              "chunk_id": "c2", "chunk_text": "text2", "vector_score": 0.88},
         ]
-        fake = _make_fake_session_factory(fake_rows)
-        monkeypatch.setattr(retrieve_mod.neo4j_client, "session", fake)
+        monkeypatch.setattr(vector_mod, "neo4j_client", _FakeNeo4jClient(fake_rows))
 
         results = _vector_search([0.2] * 1024, top_k=5)
 
@@ -205,14 +214,8 @@ class TestVectorSearch:
         assert results == []
 
     def test_handles_neo4j_exception(self, monkeypatch) -> None:
-        monkeypatch.setattr(retrieve_mod.settings, "neo4j_use_search_clause", False)
-
-        @contextmanager
-        def _failing_session():
-            raise RuntimeError("Connection failed")
-            yield  
-
-        monkeypatch.setattr(retrieve_mod.neo4j_client, "session", _failing_session)
+        monkeypatch.setattr(vector_mod.settings, "neo4j_use_search_clause", False)
+        monkeypatch.setattr(vector_mod, "neo4j_client", _FailingNeo4jClient(RuntimeError("Connection failed")))
 
         results = _vector_search([0.1] * 10, top_k=5)
         assert results == []
@@ -229,8 +232,7 @@ class TestGraphSearch:
             {"page_title": "Entity Page", "page_url": "http://ep", "page_id": "pid_ep",
              "chunk_id": "c_ent", "chunk_text": "Entity text", "graph_score": 3.5},
         ]
-        fake = _make_fake_session_factory(fake_rows)
-        monkeypatch.setattr(retrieve_mod.neo4j_client, "session", fake)
+        monkeypatch.setattr(graph_mod, "neo4j_client", _FakeNeo4jClient(fake_rows))
 
         results = _graph_search("Hà Nội", top_k=5)
 
@@ -239,19 +241,13 @@ class TestGraphSearch:
         assert results[0]["graph_score"] == 3.5
 
     def test_returns_empty_on_no_match(self, monkeypatch) -> None:
-        fake = _make_fake_session_factory([])
-        monkeypatch.setattr(retrieve_mod.neo4j_client, "session", fake)
+        monkeypatch.setattr(graph_mod, "neo4j_client", _FakeNeo4jClient([]))
 
         results = _graph_search("nonexistent entity", top_k=5)
         assert results == []
 
     def test_handles_exception(self, monkeypatch) -> None:
-        @contextmanager
-        def _failing_session():
-            raise RuntimeError("Neo4j down")
-            yield  
-
-        monkeypatch.setattr(retrieve_mod.neo4j_client, "session", _failing_session)
+        monkeypatch.setattr(graph_mod, "neo4j_client", _FailingNeo4jClient(RuntimeError("Neo4j down")))
 
         results = _graph_search("test", top_k=5)
         assert results == []
@@ -264,30 +260,26 @@ class TestGraphSearch:
 
 class TestHybridRetrieve:
     def test_combines_all_channels(self, monkeypatch) -> None:
-        """hybrid_retrieve should call BM25, vector, graph, and community channels."""
-        monkeypatch.setattr(retrieve_mod.settings, "wrrf_weight_bm25", 0.4)
-        monkeypatch.setattr(retrieve_mod.settings, "wrrf_weight_vector", 0.4)
-        monkeypatch.setattr(retrieve_mod.settings, "wrrf_weight_graph", 0.2)
-        monkeypatch.setattr(retrieve_mod.settings, "wrrf_weight_community", 0.15)
-        monkeypatch.setattr(retrieve_mod.settings, "wrrf_k", 60)
+        monkeypatch.setattr(fusion_mod.settings, "wrrf_weight_bm25", 0.4)
+        monkeypatch.setattr(fusion_mod.settings, "wrrf_weight_vector", 0.4)
+        monkeypatch.setattr(fusion_mod.settings, "wrrf_weight_graph", 0.2)
+        monkeypatch.setattr(fusion_mod.settings, "wrrf_weight_community", 0.15)
+        monkeypatch.setattr(fusion_mod.settings, "wrrf_k", 60)
 
-        # Mock embed_texts
-        monkeypatch.setattr(retrieve_mod, "embed_texts", lambda texts: [[0.1] * 10])
+        monkeypatch.setattr(fusion_mod, "embed_texts", lambda texts: [[0.1] * 10])
 
-        # Mock individual search functions
         bm25_results = [_make_row("c1", "BM25 Page", "bm25_score", 5.0)]
         vector_results = [_make_row("c2", "Vector Page", "vector_score", 0.9)]
         graph_results = [_make_row("c3", "Graph Page", "graph_score", 3.0)]
         community_results = [_make_row("c4", "Community Page", "score", 0.8)]
 
-        monkeypatch.setattr(retrieve_mod, "_run_bm25_query", lambda q, k: bm25_results)
-        monkeypatch.setattr(retrieve_mod, "_vector_search", lambda emb, k: vector_results)
-        monkeypatch.setattr(retrieve_mod, "_graph_search", lambda q, k: graph_results)
-        monkeypatch.setattr(retrieve_mod, "_community_search", lambda q, k, query_embedding=None: community_results)
+        monkeypatch.setattr(fusion_mod, "run_bm25_query", lambda q, k: bm25_results)
+        monkeypatch.setattr(fusion_mod, "vector_search", lambda emb, k: vector_results)
+        monkeypatch.setattr(fusion_mod, "graph_search", lambda q, k: graph_results)
+        monkeypatch.setattr(fusion_mod, "community_search", lambda q, k, query_embedding=None: community_results)
 
         results = hybrid_retrieve("Thủ đô Việt Nam", top_k=10)
 
-        # All 4 chunks should appear in fused results
         chunk_ids = {r["chunk_id"] for r in results}
         assert "c1" in chunk_ids
         assert "c2" in chunk_ids
@@ -295,31 +287,30 @@ class TestHybridRetrieve:
         assert "c4" in chunk_ids
 
     def test_returns_empty_when_all_channels_empty(self, monkeypatch) -> None:
-        monkeypatch.setattr(retrieve_mod, "embed_texts", lambda texts: [[0.1] * 10])
-        monkeypatch.setattr(retrieve_mod, "_run_bm25_query", lambda q, k: [])
-        monkeypatch.setattr(retrieve_mod, "_vector_search", lambda emb, k: [])
-        monkeypatch.setattr(retrieve_mod, "_graph_search", lambda q, k: [])
-        monkeypatch.setattr(retrieve_mod, "_community_search", lambda q, k, query_embedding=None: [])
+        monkeypatch.setattr(fusion_mod, "embed_texts", lambda texts: [[0.1] * 10])
+        monkeypatch.setattr(fusion_mod, "run_bm25_query", lambda q, k: [])
+        monkeypatch.setattr(fusion_mod, "vector_search", lambda emb, k: [])
+        monkeypatch.setattr(fusion_mod, "graph_search", lambda q, k: [])
+        monkeypatch.setattr(fusion_mod, "community_search", lambda q, k, query_embedding=None: [])
 
         results = hybrid_retrieve("nothing", top_k=10)
         assert results == []
 
     def test_works_without_embedding(self, monkeypatch) -> None:
-        """If embedding fails, vector and community channels are skipped."""
-        monkeypatch.setattr(retrieve_mod.settings, "wrrf_weight_bm25", 0.4)
-        monkeypatch.setattr(retrieve_mod.settings, "wrrf_weight_graph", 0.2)
-        monkeypatch.setattr(retrieve_mod.settings, "wrrf_k", 60)
+        monkeypatch.setattr(fusion_mod.settings, "wrrf_weight_bm25", 0.4)
+        monkeypatch.setattr(fusion_mod.settings, "wrrf_weight_graph", 0.2)
+        monkeypatch.setattr(fusion_mod.settings, "wrrf_k", 60)
 
         def _fail_embed(texts):
             raise RuntimeError("Embedding service down")
 
-        monkeypatch.setattr(retrieve_mod, "embed_texts", _fail_embed)
+        monkeypatch.setattr(fusion_mod, "embed_texts", _fail_embed)
 
         bm25_results = [_make_row("c1", "BM25")]
         graph_results = [_make_row("c2", "Graph")]
 
-        monkeypatch.setattr(retrieve_mod, "_run_bm25_query", lambda q, k: bm25_results)
-        monkeypatch.setattr(retrieve_mod, "_graph_search", lambda q, k: graph_results)
+        monkeypatch.setattr(fusion_mod, "run_bm25_query", lambda q, k: bm25_results)
+        monkeypatch.setattr(fusion_mod, "graph_search", lambda q, k: graph_results)
 
         results = hybrid_retrieve("test query", top_k=10)
 
@@ -328,16 +319,16 @@ class TestHybridRetrieve:
         assert "c2" in chunk_ids
 
     def test_respects_top_k(self, monkeypatch) -> None:
-        monkeypatch.setattr(retrieve_mod.settings, "wrrf_weight_bm25", 0.4)
-        monkeypatch.setattr(retrieve_mod.settings, "wrrf_k", 60)
+        monkeypatch.setattr(fusion_mod.settings, "wrrf_weight_bm25", 0.4)
+        monkeypatch.setattr(fusion_mod.settings, "wrrf_k", 60)
 
-        monkeypatch.setattr(retrieve_mod, "embed_texts", lambda texts: [[0.1] * 10])
+        monkeypatch.setattr(fusion_mod, "embed_texts", lambda texts: [[0.1] * 10])
 
         bm25_results = [_make_row(f"c{i}", f"P{i}") for i in range(20)]
-        monkeypatch.setattr(retrieve_mod, "_run_bm25_query", lambda q, k: bm25_results)
-        monkeypatch.setattr(retrieve_mod, "_vector_search", lambda emb, k: [])
-        monkeypatch.setattr(retrieve_mod, "_graph_search", lambda q, k: [])
-        monkeypatch.setattr(retrieve_mod, "_community_search", lambda q, k, query_embedding=None: [])
+        monkeypatch.setattr(fusion_mod, "run_bm25_query", lambda q, k: bm25_results)
+        monkeypatch.setattr(fusion_mod, "vector_search", lambda emb, k: [])
+        monkeypatch.setattr(fusion_mod, "graph_search", lambda q, k: [])
+        monkeypatch.setattr(fusion_mod, "community_search", lambda q, k, query_embedding=None: [])
 
         results = hybrid_retrieve("test", top_k=5)
         assert len(results) <= 5
